backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
torch>=2.2.0
torchvision>=0.15.1
Pillow==9.5.0
python-multipart==0.0.6